# lib/api-zod/src/generated/types/ugcJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UgcJobStatus } from "./ugcJobStatus";

export interface UgcJob {
  id: number;
  productUrl: string;
  status: UgcJobStatus;
  /** Human-readable description of the current step */
  progress: string;
  productTitle?: string | null;
  productSummary?: string | null;
  script?: string | null;
  /** Final narration text used for the video */
  voiceover?: string | null;
  /** Server-relative URL to the generated 9:16 MP4 (under /api/ugc/files) */
  videoUrl?: string | null;
  thumbnailUrl?: string | null;
  referenceImageUrls: string[];
  sceneImageUrls: string[];
  durationSeconds?: number | null;
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}