# lib/api-zod/src/generated/types/ugcJobInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UgcJobInput {
  /** Public URL of the product page */
  productUrl: string;
  /**
   * Array of base64-encoded images (data URIs allowed) of the actual product
   * @minItems 1
   */
  referenceImages: string[];
}